Split scheduler runner into pipeline mixins
runner.py had grown to 2801 lines mixing fetch, prepare, send, seen and status.
Move each cluster into a mixin and keep lifecycle, run_check orchestration and
replay in runner.py (960 lines).

The DTOs move to models.py so the mixins do not import runner back; runner
re-exports them so scheduler.runner.UserFetchResult still resolves. Mixins are
used rather than collaborator objects because the tests call private methods
unbound off the class and patch them as instance attributes.

Pure move: NitterTweetScheduler keeps all 97 members with identical bodies. The
only edit is _format_push_batch_summary, a staticmethod that referenced the host
class by name; it now names the mixin that defines _tweet_batches.

Also drop the now unused daily-time and timestamp format delegates.
