+++ scheduler/runner.py
@@ -3,61 +3,25 @@ from __future__ import annotations
 import asyncio
 import copy
 import datetime as dt
-import inspect
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass, field
 from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
 
 from astrbot.api import logger
 
 try:
-    from astrbot.api.all import MessageChain
-except ImportError:
-    from astrbot.api.event import MessageChain
-
-try:
-    from astrbot.api.message_components import Plain
-except ImportError:
-    from astrbot.core.message.components import Plain
-
-try:
     from ..config import (
         config_get,
-        config_set,
-        configured_merge_tweet_threshold,
-        media_only_unavailable_reason,
         migrate_default_group_config,
         parse_config_bool,
         resolve_hide_original_when_translated,
     )
-    from ..ai import (
-        TranslationReport,
-        format_ai_tweet_summary,
-    )
-    from ..shared.group_ids import (
-        DEFAULT_GROUP_NAME,
-        GLOBAL_GROUP_ID,
-        is_default_group,
-        normalize_group_id,
-    )
-    from ..shared.media_status import (
-        MEDIA_STATUS_NO_CANDIDATE,
-        MEDIA_STATUS_POLICY_SKIPPED,
-        MEDIA_STATUS_READY,
-    )
-    from .formatting import (
-        _format_limited_values as scheduler_format_limited_values,
-        format_daily_times as scheduler_format_daily_times,
-        format_group_schedule as scheduler_format_group_schedule,
-        format_merge_threshold as scheduler_format_merge_threshold,
-        format_next_daily_time as scheduler_format_next_daily_time,
-        format_timestamp as scheduler_format_timestamp,
-    )
+    from ..shared.group_ids import GLOBAL_GROUP_ID
     from .models import (
         BatchSummaryTracker,
         PendingTweetBatch,
+        PreparedBatchResult,  # noqa: F401  拆分前定义于此，保留再导出
         ScheduledCheckResult,
-        ScheduledPushResult,
+        SchedulerTaskError,  # noqa: F401  拆分前定义于此，保留再导出
+        UserFetchResult,  # noqa: F401  拆分前定义于此，保留再导出
     )
     from .config import (
         PushTargetParseResult,
@@ -65,47 +29,27 @@ try:
         SchedulerConfigReader,
         WatchUsersInfo,
     )
+    from .runner_fetch import SchedulerFetchMixin
+    from .runner_prepare import SchedulerPrepareMixin
+    from .runner_seen import SchedulerSeenMixin
+    from .runner_send import SchedulerSendMixin
+    from .runner_status import SchedulerStatusMixin
     from ..storage import StorageAdapter
-    from ..rendering import TweetMessageRenderer
-    from ..shared import TweetItem
 except ImportError:
     from config import (
         config_get,
-        config_set,
-        configured_merge_tweet_threshold,
-        media_only_unavailable_reason,
         migrate_default_group_config,
         parse_config_bool,
         resolve_hide_original_when_translated,
     )
-    from ai import (
-        TranslationReport,
-        format_ai_tweet_summary,
-    )
-    from shared.group_ids import (
-        DEFAULT_GROUP_NAME,
-        GLOBAL_GROUP_ID,
-        is_default_group,
-        normalize_group_id,
-    )
-    from shared.media_status import (
-        MEDIA_STATUS_NO_CANDIDATE,
-        MEDIA_STATUS_POLICY_SKIPPED,
-        MEDIA_STATUS_READY,
-    )
-    from scheduler.formatting import (
-        _format_limited_values as scheduler_format_limited_values,
-        format_daily_times as scheduler_format_daily_times,
-        format_group_schedule as scheduler_format_group_schedule,
-        format_merge_threshold as scheduler_format_merge_threshold,
-        format_next_daily_time as scheduler_format_next_daily_time,
-        format_timestamp as scheduler_format_timestamp,
-    )
+    from shared.group_ids import GLOBAL_GROUP_ID
     from scheduler.models import (
         BatchSummaryTracker,
         PendingTweetBatch,
+        PreparedBatchResult,  # noqa: F401  拆分前定义于此，保留再导出
         ScheduledCheckResult,
-        ScheduledPushResult,
+        SchedulerTaskError,  # noqa: F401  拆分前定义于此，保留再导出
+        UserFetchResult,  # noqa: F401  拆分前定义于此，保留再导出
     )
     from scheduler.config import (
         PushTargetParseResult,
@@ -113,9 +57,12 @@ except ImportError:
         SchedulerConfigReader,
         WatchUsersInfo,
     )
+    from scheduler.runner_fetch import SchedulerFetchMixin
+    from scheduler.runner_prepare import SchedulerPrepareMixin
+    from scheduler.runner_seen import SchedulerSeenMixin
+    from scheduler.runner_send import SchedulerSendMixin
+    from scheduler.runner_status import SchedulerStatusMixin
     from storage import StorageAdapter
-    from rendering import TweetMessageRenderer
-    from shared import TweetItem
 
 
 try:
@@ -127,44 +74,13 @@ except ZoneInfoNotFoundError:
 POLL_SECONDS = 30
 
 
-@dataclass(slots=True)
-class SchedulerTaskError:
-    message: str
-    kind: str = ""
-
-    @classmethod
-    def from_exception(cls, exc: Exception) -> "SchedulerTaskError":
-        return cls(message=str(exc), kind=type(exc).__name__)
-
-
-@dataclass(slots=True)
-class UserFetchResult:
-    index: int
-    username: str
-    instance: str = ""
-    tweets: list[TweetItem] = field(default_factory=list)
-    scanned_status_ids: list[str] = field(default_factory=list)
-    anchor_status_ids: list[str] = field(default_factory=list)
-    latest_status_id: str = ""
-    scan_complete: bool = True
-    plain_text_filtered: int = 0
-    error: SchedulerTaskError | None = None
-    # HTML search keeps these internal counters so a tag's first RT-only
-    # response can be distinguished from a genuinely empty page.
-    retweet_filtered: int = 0
-    html_raw_item_count: int = 0
-
-
-@dataclass(slots=True)
-class PreparedBatchResult:
-    batch: PendingTweetBatch
-    translation_report: TranslationReport | None = None
-    error: SchedulerTaskError | None = None
-    media_status: str = "ready"
-    media_error: str = ""
-
-
-class NitterTweetScheduler:
+class NitterTweetScheduler(
+    SchedulerFetchMixin,
+    SchedulerPrepareMixin,
+    SchedulerSendMixin,
+    SchedulerSeenMixin,
+    SchedulerStatusMixin,
+):
     def __init__(
         self,
         owner,
@@ -996,1763 +912,6 @@ class NitterTweetScheduler:
         finally:
             if run_host_skip_started and hasattr(self.nitter, "end_run_host_skip"):
                 self.nitter.end_run_host_skip()
-
-    async def status_summary(self) -> str:
-        groups = self._schedule_groups(log_invalid_targets=False)
-        default_group = next(
-            (item for item in groups if is_default_group(item.group_id)),
-            None,
-        )
-        if not groups:
-            return "Nitter 定时检查状态\n没有可用分组。"
-
-        enabled_groups = [item for item in groups if item.enabled]
-        total_users = sum(len(item.account_keys) for item in groups)
-        total_raw_users = sum(
-            (
-                item.queries_info.raw_count
-                if item.is_tag_group
-                else item.users_info.raw_count
-            )
-            for item in groups
-        )
-        total_duplicates = sum(
-            len(
-                item.queries_info.duplicates
-                if item.is_tag_group
-                else item.users_info.duplicates
-            )
-            for item in groups
-        )
-        total_invalid_users = sum(
-            len(
-                item.queries_info.invalid_entries
-                if item.is_tag_group
-                else item.users_info.invalid_entries
-            )
-            for item in groups
-        )
-        total_targets = sum(len(item.targets) for item in groups)
-        total_invalid_targets = sum(len(item.invalid_targets) for item in groups)
-        seen_map_results = await asyncio.gather(
-            *[self._get_seen_map(item.group_id) for item in groups]
-        )
-        group_seen_counts = {
-            item.group_id: len(seen_map)
-            for item, seen_map in zip(groups, seen_map_results)
-        }
-        total_seen_users = sum(group_seen_counts.values())
-
-        lines = [
-            "Nitter 定时检查状态",
-            f"调度器: {'运行中' if self.is_running else '未运行'}",
-            f"总开关: {'已启用' if self.schedule_enabled else '已关闭'}",
-            "全局检查间隔: "
-            f"{config_get(self.config, 'check_interval_minutes', 30)} 分钟",
-            "启动立即检查: "
-            f"{'已启用' if config_get(self.config, 'check_on_startup', False) else '已关闭'}",
-            "无更新提示: "
-            f"{'已启用' if config_get(self.config, 'notify_no_updates', False) else '已关闭'}",
-            f"分组数量: {len(groups)} 个（启用 {len(enabled_groups)} 个）",
-            f"QQ 合并阈值: {self._format_merge_threshold(self._merge_tweet_threshold())}",
-            "全部分组订阅账号项: "
-            f"{total_users} 个（配置 {total_raw_users} 项，"
-            f"重复 {total_duplicates} 项，无效 {total_invalid_users} 项）",
-            f"全部分组推送目标项: {total_targets} 个（无效 {total_invalid_targets} 个）",
-            f"全部分组已记录索引: {total_seen_users} 个",
-        ]
-        if default_group is not None:
-            lines.append("默认分组详情:")
-            self._append_group_status(
-                lines,
-                default_group,
-                seen_count=group_seen_counts.get(default_group.group_id, 0),
-            )
-        if len(groups) > (1 if default_group is not None else 0):
-            lines.append("其他分组详情:")
-            for item in groups:
-                if (
-                    default_group is not None
-                    and item.group_id == default_group.group_id
-                ):
-                    continue
-                self._append_group_status(
-                    lines,
-                    item,
-                    seen_count=group_seen_counts.get(item.group_id, 0),
-                )
-        return "\n".join(lines)
-
-    def deduplicate_watch_users(self) -> WatchUsersInfo:
-        info = self._watch_users_info()
-        if not info.changed:
-            return info
-
-        groups = config_get(self.config, "tweet_groups", []) or []
-        if isinstance(groups, dict):
-            groups = [groups]
-        elif not isinstance(groups, list):
-            groups = []
-        for group in groups:
-            if not isinstance(group, dict):
-                continue
-            if (
-                normalize_group_id(group.get("group_id") or group.get("name") or "")
-                == GLOBAL_GROUP_ID
-            ):
-                group["watch_users"] = info.users
-                config_set(self.config, "tweet_groups", groups)
-                break
-        else:
-            info.save_error = "未找到默认分组配置"
-            return info
-
-        save_config = getattr(self.config, "save_config", None)
-        if not callable(save_config):
-            info.save_error = "当前配置对象不支持 save_config()"
-            return info
-
-        try:
-            save_config()
-            info.saved = True
-        except Exception as exc:
-            info.save_error = str(exc)
-            logger.warning(f"[NitterTweets] 保存去重后的关注账号失败: {exc}")
-        return info
-
-    def watch_users_info(self) -> WatchUsersInfo:
-        return self._watch_users_info()
-
-    def _should_notify_no_updates(
-        self,
-        result: ScheduledCheckResult,
-        notify_no_updates: bool | None,
-        group: ScheduleGroup,
-    ) -> bool:
-        if notify_no_updates is None:
-            notify_no_updates = group.notify_no_updates
-        return bool(notify_no_updates and result.has_visible_no_update())
-
-    async def _fetch_group_users(
-        self,
-        group: ScheduleGroup,
-        fetch_limit: int,
-        skip_plain_text: bool,
-        scan_watermarks: dict[str, list[str]],
-    ) -> list[UserFetchResult]:
-        accounts = list(group.account_keys)
-        # Tag groups always serial to protect shared HTML instances.
-        if group.is_tag_group or not self._should_use_concurrent_fetch(group):
-            results = []
-            for index, username in enumerate(accounts):
-                # Add delay between queries (except before the first one)
-                if index > 0 and group.send_user_interval > 0:
-                    await asyncio.sleep(group.send_user_interval)
-                results.append(
-                    await self._fetch_group_user(
-                        group,
-                        index,
-                        username,
-                        fetch_limit,
-                        skip_plain_text,
-                        scan_watermarks.get(username),
-                        concurrent=False,
-                    )
-                )
-            return results
-
-        semaphore = asyncio.Semaphore(group.fetch_concurrency)
-
-        async def fetch_with_limit(index: int, username: str) -> UserFetchResult:
-            async with semaphore:
-                return await self._fetch_group_user(
-                    group,
-                    index,
-                    username,
-                    fetch_limit,
-                    skip_plain_text,
-                    scan_watermarks.get(username),
-                    concurrent=True,
-                )
-
-        tasks = [
-            fetch_with_limit(index, username) for index, username in enumerate(accounts)
-        ]
-        return list(await asyncio.gather(*tasks))
-
-    async def _fetch_group_user(
-        self,
-        group: ScheduleGroup,
-        index: int,
-        username: str,
-        fetch_limit: int,
-        skip_plain_text: bool,
-        scan_watermark: list[str] | None,
-        *,
-        concurrent: bool,
-    ) -> UserFetchResult:
-        if group.is_tag_group:
-            return await self._fetch_group_query(
-                group,
-                index,
-                username,
-                fetch_limit,
-                skip_plain_text=skip_plain_text,
-            )
-        try:
-            scheduler_method = (
-                "fetch_tweets_for_scheduler_from_instances"
-                if concurrent
-                else "fetch_tweets_for_scheduler"
-            )
-            fetch_for_scheduler = getattr(self.nitter, scheduler_method, None)
-            if callable(fetch_for_scheduler):
-                if concurrent:
-                    instance, scan_result = await fetch_for_scheduler(
-                        username,
-                        scan_watermark,
-                        group.concurrent_fetch_instances,
-                        start_index=index,
-                        skip_plain_text=skip_plain_text,
-                        retry_attempts=getattr(
-                            self.nitter, "retry_attempts", 2
-                        ),
-                    )
-                else:
-                    instance, scan_result = await fetch_for_scheduler(
-                        username,
-                        scan_watermark,
-                        skip_plain_text=skip_plain_text,
-                    )
-                raw_anchor_status_ids = getattr(scan_result, "anchor_status_ids", None)
-                anchor_status_ids = (
-                    list(scan_result.scanned_status_ids)[:20]
-                    if raw_anchor_status_ids is None
-                    else list(raw_anchor_status_ids)
-                )
-                tweets = list(scan_result.tweets)
-                if not tweets and self._user_html_fallback_enabled():
-                    html_result = await self._fetch_user_html_fallback(
-                        index,
-                        username,
-                        fetch_limit,
-                        skip_plain_text=skip_plain_text,
-                    )
-                    if html_result is not None:
-                        return html_result
-                return UserFetchResult(
-                    index=index,
-                    username=username,
-                    instance=instance,
-                    tweets=tweets,
-                    scanned_status_ids=list(scan_result.scanned_status_ids),
-                    anchor_status_ids=anchor_status_ids,
-                    latest_status_id=str(scan_result.latest_status_id or ""),
-                    scan_complete=bool(scan_result.complete),
-                    plain_text_filtered=int(scan_result.plain_text_filtered or 0),
-                )
-
-            if concurrent:
-                (
-                    instance,
-                    tweets,
-                    plain_text_filtered,
-                ) = await self.nitter.fetch_tweets_with_stats_from_instances(
-                    username,
-                    fetch_limit,
-                    group.concurrent_fetch_instances,
-                    start_index=index,
-                    skip_plain_text=skip_plain_text,
-                    retry_attempts=getattr(self.nitter, "retry_attempts", 2),
-                )
-            else:
-                (
-                    instance,
-                    tweets,
-                    plain_text_filtered,
-                ) = await self.nitter.fetch_tweets_with_stats(
-                    username, fetch_limit, skip_plain_text=skip_plain_text
-                )
-            if not tweets and self._user_html_fallback_enabled():
-                html_result = await self._fetch_user_html_fallback(
-                    index,
-                    username,
-                    fetch_limit,
-                    skip_plain_text=skip_plain_text,
-                )
-                if html_result is not None:
-                    return html_result
-        except Exception as exc:
-            if self._user_html_fallback_enabled():
-                html_result = await self._fetch_user_html_fallback(
-                    index,
-                    username,
-                    fetch_limit,
-                    skip_plain_text=skip_plain_text,
-                )
-                if html_result is not None:
-                    return html_result
-            return UserFetchResult(
-                index=index,
-                username=username,
-                error=SchedulerTaskError.from_exception(exc),
-            )
-        return UserFetchResult(
-            index=index,
-            username=username,
-            instance=instance,
-            tweets=tweets,
-            scanned_status_ids=[tweet.status_id for tweet in tweets if tweet.status_id],
-            anchor_status_ids=[
-                tweet.status_id for tweet in tweets[:20] if tweet.status_id
-            ],
-            latest_status_id=(tweets[0].status_id if tweets else ""),
-            plain_text_filtered=plain_text_filtered,
-        )
-
-    def _user_html_fallback_enabled(self) -> bool:
-        if self.html_backend is None:
-            return False
-        return parse_config_bool(
-            config_get(self.config, "user_html_fallback", False), False
-        )
-
-    async def _fetch_user_html_fallback(
-        self,
-        index: int,
-        username: str,
-        fetch_limit: int,
-        *,
-        skip_plain_text: bool = False,
-    ) -> UserFetchResult | None:
-        if self.html_backend is None:
-            return None
-        try:
-            instance, tweets = await asyncio.to_thread(
-                self.html_backend.fetch_user, username, fetch_limit
-            )
-        except Exception as exc:
-            logger.warning(
-                f"[NitterTweets] HTML 用户页回退失败: @{username}, error={exc}"
-            )
-            return None
-        if not tweets:
-            return None
-        tweets, plain_text_filtered = self._filter_html_tweets_plain_text(
-            list(tweets), skip_plain_text=skip_plain_text
-        )
-        return UserFetchResult(
-            index=index,
-            username=username,
-            instance=instance,
-            tweets=tweets,
-            scanned_status_ids=[tweet.status_id for tweet in tweets if tweet.status_id],
-            anchor_status_ids=[
-                tweet.status_id for tweet in tweets[:20] if tweet.status_id
-            ],
-            latest_status_id=(tweets[0].status_id if tweets else ""),
-            scan_complete=True,
-            plain_text_filtered=plain_text_filtered,
-        )
-
-    @staticmethod
-    def _filter_html_tweets_plain_text(
-        tweets: list[TweetItem],
-        *,
-        skip_plain_text: bool,
-    ) -> tuple[list[TweetItem], int]:
-        """HTML items already carry author media; drop pure-text when filtering."""
-        if not skip_plain_text or not tweets:
-            return tweets, 0
-        kept = [tweet for tweet in tweets if tweet.media]
-        return kept, len(tweets) - len(kept)
-
-    async def _fetch_group_query(
-        self,
-        group: ScheduleGroup,
-        index: int,
-        account_key: str,
-        fetch_limit: int,
-        *,
-        skip_plain_text: bool = False,
-    ) -> UserFetchResult:
-        query_item = next(
-            (item for item in group.queries if item.account_key == account_key),
-            None,
-        )
-        if query_item is None:
-            return UserFetchResult(
-                index=index,
-                username=account_key,
-                error=SchedulerTaskError.from_exception(
-                    RuntimeError(f"missing watch query for {account_key}")
-                ),
-            )
-        if self.html_backend is None:
-            return UserFetchResult(
-                index=index,
-                username=account_key,
-                error=SchedulerTaskError.from_exception(
-                    RuntimeError("html_backend unavailable")
-                ),
-            )
-        html_backend = self.html_backend
-
-        logger.info(
-            f"[NitterTweets] 标签查询开始: group={group.group_id}, "
-            f"query={query_item.query}, type={query_item.type}, limit={fetch_limit}"
-        )
-
-        try:
-            instance, tweets = await asyncio.to_thread(
-                lambda: html_backend.search(
-                    query_item.query, fetch_limit, kind=query_item.type
-                )
-            )
-            logger.info(
-                f"[NitterTweets] 标签查询成功: group={group.group_id}, "
-                f"query={query_item.query}, instance={instance}, "
-                f"tweets={len(list(tweets))}"
-            )
-        except Exception as exc:
-            logger.warning(
-                f"[NitterTweets] 标签查询失败: group={group.group_id}, "
-                f"query={query_item.query}, error={type(exc).__name__}: {exc}"
-            )
-            return UserFetchResult(
-                index=index,
-                username=account_key,
-                error=SchedulerTaskError.from_exception(exc),
-            )
-        retweet_filtered = max(0, int(getattr(tweets, "retweet_filtered", 0) or 0))
-        html_raw_item_count = max(
-            0, int(getattr(tweets, "raw_item_count", 0) or 0)
-        )
-        tweets, plain_text_filtered = self._filter_html_tweets_plain_text(
-            list(tweets), skip_plain_text=skip_plain_text
-        )
-        return UserFetchResult(
-            index=index,
-            username=account_key,
-            instance=instance,
-            tweets=tweets,
-            scanned_status_ids=[tweet.status_id for tweet in tweets if tweet.status_id],
-            anchor_status_ids=[
-                tweet.status_id for tweet in tweets[:20] if tweet.status_id
-            ],
-            latest_status_id=(tweets[0].status_id if tweets else ""),
-            scan_complete=True,
-            plain_text_filtered=plain_text_filtered,
-            retweet_filtered=retweet_filtered,
-            html_raw_item_count=html_raw_item_count,
-        )
-
-    @staticmethod
-    def _should_use_concurrent_fetch(group: ScheduleGroup) -> bool:
-        return (
-            bool(group.concurrent_fetch_enabled)
-            and bool(group.concurrent_fetch_instances)
-            and group.fetch_concurrency > 1
-        )
-
-    @staticmethod
-    def _should_use_concurrent_prepare(group: ScheduleGroup) -> bool:
-        return bool(group.concurrent_prepare_enabled) and group.prepare_concurrency > 1
-
-    async def _send_no_update_notice(
-        self, result: ScheduledCheckResult, target_interval: float
-    ) -> None:
-        text = result.format_message("Nitter 定时检查无更新")
-        success = 0
-        for target_index, umo in enumerate(result.targets):
-            try:
-                sent = await self.context.send_message(umo, MessageChain([Plain(text)]))
-                if sent is not False:
-                    success += 1
-                else:
-                    logger.warning(
-                        "[NitterTweets] 发送无更新提示失败: "
-                        f"target={umo}, error=未找到目标平台或平台不支持主动发送"
-                    )
-            except Exception as exc:
-                logger.warning(
-                    f"[NitterTweets] 发送无更新提示失败: target={umo}, error={exc}"
-                )
-            if target_index < len(result.targets) - 1 and target_interval > 0:
-                await asyncio.sleep(target_interval)
-        self._log_verbose_info(
-            f"[NitterTweets] 无更新提示发送完成: success={success}/{len(result.targets)}"
-        )
-
-    @staticmethod
-    def _set_discovered_batch_progress(
-        discovered_batches: list[PendingTweetBatch],
-    ) -> None:
-        account_total = len(discovered_batches)
-        for account_index, discovered_batch in enumerate(discovered_batches, 1):
-            discovered_batch.account_index = account_index
-            discovered_batch.account_total = account_total
-
-    async def _prepare_discovered_batches_serial(
-        self,
-        group: ScheduleGroup,
-        discovered_batches: list[PendingTweetBatch],
-        result: ScheduledCheckResult,
-        target_umo: str,
-        seen_map: dict[str, list[str]],
-        immediate_targets: list[str],
-        buffered_targets: list[str],
-        target_interval: float,
-        user_interval: float,
-        group_label: str,
-        immediate_batch_summary_tracker: BatchSummaryTracker,
-        immediate_batches_sent: int,
-    ) -> tuple[list[PendingTweetBatch], int]:
-        pending_batches: list[PendingTweetBatch] = []
-        current_batch: PendingTweetBatch | None = None
-        try:
-            for discovered_batch in discovered_batches:
-                username = discovered_batch.username
-                new_tweets = discovered_batch.tweets
-                prepared_count = 0
-                for tweet_index, tweet in enumerate(new_tweets, 1):
-                    current_batch = self._single_tweet_batch(
-                        discovered_batch,
-                        tweet,
-                        tweet_index,
-                        seen_map,
-                    )
-                    try:
-                        prepared = await self._prepare_immediate_batch(
-                            current_batch, target_umo
-                        )
-                    except Exception as exc:
-                        await self._record_prepare_failure(result, current_batch, exc)
-                        current_batch = None
-                        continue
-
-                    if prepared.media_status != MEDIA_STATUS_READY:
-                        await self._handle_media_prepare_status(
-                            result,
-                            current_batch,
-                            prepared.media_status,
-                            prepared.media_error,
-                            seen_map,
-                            group.group_id,
-                        )
-                        current_batch = None
-                        continue
-
-                    await self._handle_immediate_prepare_success(
-                        group, prepared, seen_map
-                    )
-                    prepared_count += 1
-                    (
-                        pending_batches,
-                        immediate_batches_sent,
-                    ) = await self._send_or_buffer_immediate_batch(
-                        current_batch,
-                        group,
-                        pending_batches,
-                        result,
-                        immediate_targets,
-                        buffered_targets,
-                        target_interval,
-                        user_interval,
-                        group_label,
-                        immediate_batch_summary_tracker,
-                        immediate_batches_sent,
-                        seen_map,
-                        batch_progress=(tweet_index, len(new_tweets)),
-                    )
-                    current_batch = None
-                self._log_prepare_progress(username, prepared_count, len(new_tweets))
-        except BaseException:
-            batches_to_clean = list(pending_batches)
-            if current_batch is not None:
-                batches_to_clean.append(current_batch)
-            await self._cleanup_batch_media_many(batches_to_clean)
-            raise
-        return pending_batches, immediate_batches_sent
-
-    async def _prepare_batches_concurrently(
-        self,
-        batches: list[PendingTweetBatch],
-        concurrency: int,
-        prepare_one: Callable[[PendingTweetBatch], Awaitable[PreparedBatchResult]],
-        on_result: Callable[[PreparedBatchResult], Awaitable[None]] | None = None,
-    ) -> list[PreparedBatchResult]:
-        semaphore = asyncio.Semaphore(concurrency)
-
-        async def prepare(batch: PendingTweetBatch) -> PreparedBatchResult:
-            async with semaphore:
-                try:
-                    return await prepare_one(batch)
-                except Exception as exc:
-                    return PreparedBatchResult(
-                        batch=batch,
-                        error=SchedulerTaskError.from_exception(exc),
-                    )
-
-        tasks = [asyncio.create_task(prepare(batch)) for batch in batches]
-        results: list[PreparedBatchResult] = []
-        try:
-            for task in asyncio.as_completed(tasks):
-                prepared = await task
-                results.append(prepared)
-                if on_result is not None:
-                    await on_result(prepared)
-        except BaseException:
-            for task in tasks:
-                if not task.done():
-                    task.cancel()
-            await asyncio.gather(*tasks, return_exceptions=True)
-            raise
-        return results
-
-    async def _prepare_immediate_batches_concurrently(
-        self,
-        group: ScheduleGroup,
-        discovered_batches: list[PendingTweetBatch],
-        result: ScheduledCheckResult,
-        target_umo: str,
-        seen_map: dict[str, list[str]],
-        immediate_targets: list[str],
-        buffered_targets: list[str],
-        target_interval: float,
-        user_interval: float,
-        group_label: str,
-        immediate_batch_summary_tracker: BatchSummaryTracker,
-        immediate_batches_sent: int,
-    ) -> tuple[list[PendingTweetBatch], int]:
-        single_batches: list[PendingTweetBatch] = []
-        for discovered_batch in discovered_batches:
-            for tweet_index, tweet in enumerate(discovered_batch.tweets, 1):
-                single_batches.append(
-                    self._single_tweet_batch(
-                        discovered_batch,
-                        tweet,
-                        tweet_index,
-                        seen_map,
-                    )
-                )
-
-        pending_batches: list[PendingTweetBatch] = []
-        prepared_count_by_user: dict[str, int] = {}
-        total_by_user: dict[str, int] = {}
-        for batch in discovered_batches:
-            total_by_user[batch.username] = total_by_user.get(batch.username, 0) + len(
-                batch.tweets
-            )
-
-        async def consume(prepared: PreparedBatchResult) -> None:
-            nonlocal pending_batches, immediate_batches_sent
-            batch = prepared.batch
-            tweet_index = batch.tweet_index
-            if prepared.error:
-                await self._record_prepare_failure(result, batch, prepared.error)
-                return
-            if prepared.media_status != MEDIA_STATUS_READY:
-                await self._handle_media_prepare_status(
-                    result,
-                    batch,
-                    prepared.media_status,
-                    prepared.media_error,
-                    seen_map,
-                    group.group_id,
-                )
-                return
-
-            await self._handle_immediate_prepare_success(group, prepared, seen_map)
-            prepared_count_by_user[batch.username] = (
-                prepared_count_by_user.get(batch.username, 0) + 1
-            )
-            (
-                pending_batches,
-                immediate_batches_sent,
-            ) = await self._send_or_buffer_immediate_batch(
-                batch,
-                group,
-                pending_batches,
-                result,
-                immediate_targets,
-                buffered_targets,
-                target_interval,
-                user_interval,
-                group_label,
-                immediate_batch_summary_tracker,
-                immediate_batches_sent,
-                seen_map,
-                batch_progress=(tweet_index, batch.tweet_total),
-            )
-
-        try:
-            await self._prepare_batches_concurrently(
-                single_batches,
-                group.prepare_concurrency,
-                lambda batch: self._prepare_immediate_batch(batch, target_umo),
-                on_result=consume,
-            )
-        except BaseException:
-            await self._cleanup_batch_media_many([*single_batches, *pending_batches])
-            raise
-
-        for discovered_batch in discovered_batches:
-            prepared_count = prepared_count_by_user.get(discovered_batch.username, 0)
-            self._log_prepare_progress(
-                discovered_batch.username,
-                prepared_count,
-                total_by_user.get(discovered_batch.username, 0),
-            )
-        return pending_batches, immediate_batches_sent
-
-    async def _prepare_immediate_batch(
-        self,
-        batch: PendingTweetBatch,
-        target_umo: str,
-    ) -> PreparedBatchResult:
-        translation_report = None
-        if not batch.media_only:
-            translation_report = await self.translator.attach_translations(
-                batch.tweets, target_umo
-            )
-
-        media_status = MEDIA_STATUS_READY
-        media_error = ""
-        if batch.media_only:
-            media_results_method = getattr(
-                self.media, "attach_media_with_results", None
-            )
-            if callable(media_results_method):
-                reports = await media_results_method(batch.tweets)
-                report = reports[0] if reports else None
-                media_status = str(getattr(report, "status", MEDIA_STATUS_NO_CANDIDATE))
-                media_error = str(getattr(report, "error", "") or "")
-            else:
-                await self.media.attach_media(batch.tweets)
-                media_status = (
-                    MEDIA_STATUS_READY
-                    if self._has_prepared_media(batch.tweets)
-                    else MEDIA_STATUS_NO_CANDIDATE
-                )
-        else:
-            await self.media.attach_media(batch.tweets)
-        return PreparedBatchResult(
-            batch=batch,
-            translation_report=translation_report,
-            media_status=media_status,
-            media_error=media_error,
-        )
-
-    @staticmethod
-    def _single_tweet_batch(
-        discovered_batch: PendingTweetBatch,
-        tweet: TweetItem,
-        tweet_index: int,
-        seen_map: dict[str, list[str]],
-    ) -> PendingTweetBatch:
-        return PendingTweetBatch(
-            username=discovered_batch.username,
-            instance=discovered_batch.instance,
-            tweets=[tweet],
-            fetched_ids=[tweet.status_id] if tweet.status_id else [],
-            seen_ids=seen_map.get(discovered_batch.username, []),
-            account_index=discovered_batch.account_index,
-            account_total=discovered_batch.account_total,
-            tweet_index=tweet_index,
-            tweet_total=len(discovered_batch.tweets),
-            media_only=discovered_batch.media_only,
-            omit_status_url=bool(getattr(discovered_batch, 'omit_status_url', True)),
-            hide_original_when_translated=bool(getattr(discovered_batch, "hide_original_when_translated", False)),
-        )
-
-    async def _record_prepare_failure(
-        self,
-        result: ScheduledCheckResult,
-        batch: PendingTweetBatch,
-        error: SchedulerTaskError | Exception,
-    ) -> None:
-        username = batch.username
-        tweet = batch.tweets[0]
-        tweet_index = batch.tweet_index
-        status_id = tweet.status_id or f"index-{tweet_index}"
-        if isinstance(error, SchedulerTaskError):
-            error_message = error.message
-        else:
-            error_message = str(error)
-        await self._cleanup_batch_media(batch)
-        result.failed_users[f"{username}:{status_id}"] = (
-            f"推文准备失败: {error_message}"
-        )
-        logger.warning(
-            "[NitterTweets] 定时推送准备失败: "
-            f"username={username}, status={status_id}, error={error_message}"
-        )
-
-    async def _handle_immediate_prepare_success(
-        self,
-        group: ScheduleGroup,
-        prepared: PreparedBatchResult,
-        seen_map: dict[str, list[str]],
-    ) -> None:
-        batch = prepared.batch
-        if batch.media_only:
-            return
-        self._log_ai_process_results(
-            batch.username,
-            batch.tweets,
-            prepared.translation_report,
-            progress_index=batch.tweet_index,
-            progress_total=batch.tweet_total,
-        )
-
-    def _log_prepare_progress(
-        self,
-        username: str,
-        prepared_count: int,
-        total_count: int,
-    ) -> None:
-        self._log_verbose_info(
-            f"[NitterTweets] prepared @{username} {prepared_count}/"
-            f"{total_count} new tweets for scheduled push"
-        )
-
-    async def _send_or_buffer_immediate_batch(
-        self,
-        batch: PendingTweetBatch,
-        group: ScheduleGroup,
-        pending_batches: list[PendingTweetBatch],
-        result: ScheduledCheckResult,
-        immediate_targets: list[str],
-        buffered_targets: list[str],
-        target_interval: float,
-        user_interval: float,
-        group_label: str,
-        immediate_batch_summary_tracker: BatchSummaryTracker,
-        immediate_batches_sent: int,
-        seen_map: dict[str, list[str]],
-        *,
-        batch_progress: tuple[int, int],
-    ) -> tuple[list[PendingTweetBatch], int]:
-        if buffered_targets:
-            try:
-                if immediate_targets:
-                    if immediate_batches_sent > 0 and user_interval > 0:
-                        await asyncio.sleep(user_interval)
-                    await self._send_per_user_updates(
-                        [batch],
-                        result,
-                        immediate_targets,
-                        target_interval,
-                        0.0,
-                        group_label=group_label,
-                        batch_summary_tracker=immediate_batch_summary_tracker,
-                        batch_progress=batch_progress,
-                        history_group_id=group.group_id,
-                        history_source="scheduled",
-                    )
-                    immediate_batches_sent += 1
-                pending_batches.append(batch)
-            except BaseException:
-                await self._cleanup_batch_media(batch)
-                raise
-            return pending_batches, immediate_batches_sent
-
-        try:
-            if immediate_targets:
-                if immediate_batches_sent > 0 and user_interval > 0:
-                    await asyncio.sleep(user_interval)
-                await self._send_per_user_updates(
-                    [batch],
-                    result,
-                    immediate_targets,
-                    target_interval,
-                    0.0,
-                    group_label=group_label,
-                    batch_summary_tracker=immediate_batch_summary_tracker,
-                    batch_progress=batch_progress,
-                    history_group_id=group.group_id,
-                    history_source="scheduled",
-                )
-                if (
-                    self._all_targets_delivered(immediate_targets, batch)
-                    and batch.tweets[0].status_id
-                ):
-                    await self._store_incremental_seen_ids(
-                        group.group_id,
-                        batch.username,
-                        [batch.tweets[0].status_id],
-                        seen_map,
-                    )
-                immediate_batches_sent += 1
-        finally:
-            await self._cleanup_batch_media(batch)
-        return pending_batches, immediate_batches_sent
-
-    async def _cleanup_batch_media(self, batch: PendingTweetBatch) -> None:
-        """Clean one prepared batch at most once, including cancellation paths."""
-        if batch.media_cleaned:
-            return
-        batch.media_cleaned = True
-        try:
-            await asyncio.to_thread(self.media.cleanup_after_send, batch.tweets)
-        except BaseException:
-            batch.media_cleaned = False
-            raise
-
-    async def _cleanup_batch_media_many(self, batches: list[PendingTweetBatch]) -> None:
-        for batch in batches:
-            try:
-                await self._cleanup_batch_media(batch)
-            except Exception as exc:
-                logger.warning(
-                    "[NitterTweets] 批量媒体清理失败，继续清理其余批次: "
-                    f"user=@{batch.username}, error={exc}"
-                )
-
-    async def _send_per_user_updates(
-        self,
-        batches,
-        result: ScheduledCheckResult,
-        targets: list[str],
-        target_interval: float,
-        user_interval: float,
-        merge_existing_stats: bool = False,
-        group_label: str = "",
-        batch_summary: str = "",
-        batch_summary_tracker: BatchSummaryTracker | None = None,
-        batch_progress: tuple[int, int] | None = None,
-        on_target_delivered=None,
-        history_group_id: str = "",
-        history_source: str = "scheduled",
-    ) -> int:
-        if batch_summary and batch_summary_tracker is None:
-            batch_summary_tracker = BatchSummaryTracker(batch_summary)
-
-        total_success = 0
-        for batch_index, batch in enumerate(batches):
-            success = 0
-            attempted = 0
-            for target_index, umo in enumerate(targets):
-                if umo in batch.delivered_targets:
-                    continue
-                attempted += 1
-                try:
-                    header_text = self._scheduled_update_header(batch, batch_progress)
-                    target_batch_summary = (
-                        batch_summary_tracker.for_target(umo)
-                        if batch_summary_tracker is not None
-                        else ""
-                    )
-                    if target_batch_summary:
-                        summary_outcome = await self.sender.send_summary_to_umo(
-                            self.context,
-                            umo,
-                            target_batch_summary,
-                        )
-                        if not summary_outcome.success:
-                            logger.warning(
-                                "[NitterTweets] 定时推送概括发送失败: "
-                                f"target={umo}, error={summary_outcome.error}"
-                            )
-                        if summary_outcome.warning:
-                            result.delivery_warnings.append(summary_outcome.warning)
-                        if batch_summary_tracker is not None:
-                            # A target receives at most one standalone summary;
-                            # a failed attempt must not be repeated before each tweet.
-                            batch_summary_tracker.mark_delivered(umo)
-                    tweet_start_index = self._scheduled_tweet_start_index(
-                        batch,
-                        batch_progress,
-                    )
-                    send_kwargs = {
-                        # omit_status_url/link_style filled below
-
-                        "group_label": group_label,
-                        "header_text": header_text,
-                        "batch_summary": "",
-                        "tweet_start_index": tweet_start_index,
-                    }
-                    if batch.media_only:
-                        send_kwargs["media_only"] = True
-                    send_kwargs["omit_status_url"] = bool(getattr(batch, "omit_status_url", True))
-                    send_kwargs["hide_original_when_translated"] = bool(getattr(batch, "hide_original_when_translated", False))
-                    outcome = await self.sender.send_to_umo_with_outcome(
-                        self.context,
-                        umo,
-                        batch.username,
-                        batch.instance,
-                        batch.tweets,
-                        **send_kwargs,
-                    )
-                    delivery_complete = self._delivery_is_complete(outcome)
-                    if delivery_complete:
-                        await self._mark_batch_target_delivered(
-                            batch, umo, on_target_delivered
-                        )
-                    history_status_ids = self._delivery_history_status_ids(
-                        outcome, batch.tweets
-                    )
-                    if history_status_ids:
-                        await self._record_batch_push_history(
-                            history_group_id,
-                            batch,
-                            umo,
-                            history_source,
-                            delivery_status=getattr(
-                                outcome, "delivery_status", "success"
-                            ),
-                            delivery_error=getattr(outcome, "delivery_error", ""),
-                            status_ids=history_status_ids,
-                        )
-                    if delivery_complete:
-                        success += 1
-                    if outcome.warning:
-                        result.delivery_warnings.append(outcome.warning)
-                except Exception as exc:
-                    logger.warning(
-                        f"[NitterTweets] 定时推送失败: username={batch.username}, "
-                        f"target={umo}, error={exc}"
-                    )
-                if target_index < len(targets) - 1 and target_interval > 0:
-                    await asyncio.sleep(target_interval)
-            self._record_scheduled_push(
-                result,
-                batch,
-                success,
-                attempted,
-                merge_existing_stats=merge_existing_stats,
-            )
-            total_success += success
-            if batch_progress:
-                progress_text = f" progress={batch_progress[0]}/{batch_progress[1]}"
-            elif len(batches) > 1:
-                progress_text = f" progress={batch_index + 1}/{len(batches)}"
-            else:
-                progress_text = ""
-            self._log_verbose_info(
-                f"[NitterTweets] 推送完成{progress_text}: "
-                f"username={batch.username}, tweets={len(batch.tweets)}, "
-                f"targets={success}/{len(targets)}"
-            )
-            if batch_index < len(batches) - 1 and user_interval > 0:
-                await asyncio.sleep(user_interval)
-        return total_success
-
-    @staticmethod
-    def _scheduled_update_header(
-        batch: PendingTweetBatch, batch_progress: tuple[int, int] | None = None
-    ) -> str:
-        del batch, batch_progress
-        return ""
-
-    @staticmethod
-    def _scheduled_tweet_start_index(
-        batch: PendingTweetBatch,
-        batch_progress: tuple[int, int] | None = None,
-    ) -> int:
-        if batch_progress and batch_progress[0] > 0:
-            return batch_progress[0]
-        if batch.tweet_index > 0:
-            return max(batch.tweet_index - len(batch.tweets) + 1, 1)
-        return 1
-
-    def _log_ai_process_results(
-        self,
-        username: str,
-        tweets,
-        translation_report=None,
-        progress_index: int = 0,
-        progress_total: int = 0,
-    ) -> None:
-        total = progress_total or len(tweets)
-        start = progress_index or 1
-        if self.brief_log_enabled:
-            return
-        for offset, tweet in enumerate(tweets):
-            logger.info(
-                format_ai_tweet_summary(
-                    username,
-                    tweet,
-                    translation_report,
-                    start + offset,
-                    total,
-                )
-            )
-
-    def _record_scheduled_push(
-        self,
-        result: ScheduledCheckResult,
-        batch: PendingTweetBatch,
-        success_targets: int,
-        total_targets: int,
-        *,
-        merge_existing_stats: bool = False,
-    ) -> None:
-        batch_key = self._scheduled_batch_key(batch)
-        if merge_existing_stats:
-            for push in result.pushes:
-                if push.batch_key == batch_key:
-                    push.success_targets += success_targets
-                    push.total_targets += total_targets
-                    return
-        result.pushes.append(
-            ScheduledPushResult(
-                username=batch.username,
-                new_count=len(batch.tweets),
-                success_targets=success_targets,
-                total_targets=total_targets,
-                batch_key=batch_key,
-            )
-        )
-
-    @staticmethod
-    def _scheduled_batch_key(batch: PendingTweetBatch) -> str:
-        status_ids = tuple(
-            str(getattr(tweet, "status_id", "")) for tweet in batch.tweets
-        )
-        return repr((batch.username, batch.instance, status_ids))
-
-    async def _record_batch_push_history(
-        self,
-        group_id: str,
-        batch: PendingTweetBatch,
-        target_umo: str,
-        source: str,
-        delivery_status: str = "success",
-        delivery_error: str = "",
-        status_ids: tuple[str, ...] | None = None,
-    ) -> None:
-        if not group_id:
-            return
-        selected_status_ids = None if status_ids is None else set(status_ids)
-        for tweet in batch.tweets:
-            if not getattr(tweet, "status_id", ""):
-                continue
-            if (
-                selected_status_ids is not None
-                and str(tweet.status_id) not in selected_status_ids
-            ):
-                continue
-            try:
-                await self.storage.record_push_history(
-                    group_id,
-                    batch.username,
-                    tweet,
-                    target_umo,
-                    source,
-                    batch.instance,
-                    delivery_status,
-                    delivery_error,
-                )
-            except Exception as exc:
-                logger.warning(
-                    "[NitterTweets] 记录推送历史失败: "
-                    f"group={group_id}, username={batch.username}, "
-                    f"status={tweet.status_id}, target={target_umo}, error={exc}"
-                )
-
-    async def _send_merged_updates(
-        self,
-        batches: list[PendingTweetBatch],
-        result: ScheduledCheckResult,
-        targets: list[str],
-        target_interval: float,
-        group_label: str = "",
-        batch_summary: str = "",
-        on_target_delivered=None,
-        history_group_id: str = "",
-        history_source: str = "scheduled",
-    ) -> None:
-        success = 0
-        attempts = 0
-        for target_index, umo in enumerate(targets):
-            target_batches = [
-                batch for batch in batches if umo not in batch.delivered_targets
-            ]
-            if not target_batches:
-                continue
-            attempts += 1
-            try:
-                merge_kwargs = {
-                    "omit_status_url": all(
-                        bool(getattr(batch, "omit_status_url", True))
-                        for batch in target_batches
-                    ),
-                    "hide_original_when_translated": all(
-                        bool(getattr(batch, "hide_original_when_translated", False))
-                        for batch in target_batches
-                    ),
-                }
-                if any(batch.media_only for batch in target_batches):
-                    merge_kwargs["media_only"] = True
-                outcome = await self.sender.send_merged_to_umo(
-                    self.context,
-                    umo,
-                    self._tweet_batches(target_batches),
-                    group_label=group_label,
-                    batch_summary=batch_summary,
-                    **merge_kwargs,
-                )
-                delivery_complete = self._delivery_is_complete(outcome)
-                if delivery_complete:
-                    for batch in target_batches:
-                        await self._mark_batch_target_delivered(
-                            batch, umo, on_target_delivered
-                        )
-                history_status_ids = self._delivery_history_status_ids(
-                    outcome,
-                    [
-                        tweet
-                        for batch in target_batches
-                        for tweet in batch.tweets
-                    ],
-                )
-                if history_status_ids:
-                    selected_status_ids = set(history_status_ids)
-                    for batch in target_batches:
-                        batch_status_ids = tuple(
-                            str(tweet.status_id)
-                            for tweet in batch.tweets
-                            if tweet.status_id
-                            and str(tweet.status_id) in selected_status_ids
-                        )
-                        if not batch_status_ids:
-                            continue
-                        await self._record_batch_push_history(
-                            history_group_id,
-                            batch,
-                            umo,
-                            history_source,
-                            delivery_status=getattr(
-                                outcome, "delivery_status", "success"
-                            ),
-                            delivery_error=getattr(outcome, "delivery_error", ""),
-                            status_ids=batch_status_ids,
-                        )
-                if delivery_complete:
-                    success += 1
-                if outcome.success:
-                    if outcome.warning:
-                        result.delivery_warnings.append(outcome.warning)
-                    if outcome.mode not in {
-                        "full_forward",
-                        "raw_forward",
-                        "forward_without_videos",
-                        "raw_forward_without_videos",
-                        "uncertain_delivery",
-                    }:
-                        self._log_verbose_info(
-                            f"[NitterTweets] QQ 合并推送使用普通发送路径：mode={outcome.mode}"
-                        )
-                else:
-                    logger.warning(
-                        f"[NitterTweets] 定时合并推送失败: target={umo}, "
-                        f"error={outcome.error}"
-                    )
-            except Exception as exc:
-                logger.warning(
-                    f"[NitterTweets] 定时合并推送失败: target={umo}, error={exc}"
-                )
-            if target_index < len(targets) - 1 and target_interval > 0:
-                await asyncio.sleep(target_interval)
-
-        result.merged_push_success_targets = success
-        result.merged_push_total_targets = attempts
-        self._log_verbose_info(
-            f"[NitterTweets] 合并推送完成: tweets={result.new_tweet_count}, "
-            f"users={len(batches)}, qq_targets={success}/{attempts}"
-        )
-
-    async def _send_prepared_batches(
-        self,
-        batches: list[PendingTweetBatch],
-        result: ScheduledCheckResult,
-        targets: list[str],
-        target_interval: float,
-        user_interval: float,
-        record_merge_placeholders: bool = True,
-        merge_existing_stats: bool = False,
-        group_label: str = "",
-        batch_summary: str = "",
-        on_target_delivered=None,
-        history_group_id: str = "",
-        history_source: str = "scheduled",
-    ) -> None:
-        if self._should_merge_batches(batches, result.merge_tweet_threshold):
-            merge_targets, ordinary_targets = self._split_merge_targets(targets)
-        else:
-            merge_targets, ordinary_targets = [], targets
-
-        if merge_targets:
-            result.push_mode = (
-                "mixed" if ordinary_targets or result.pushes else "merged"
-            )
-            if ordinary_targets:
-                await self._send_per_user_updates(
-                    batches,
-                    result,
-                    ordinary_targets,
-                    target_interval,
-                    user_interval,
-                    merge_existing_stats=merge_existing_stats,
-                    group_label=group_label,
-                    batch_summary=batch_summary,
-                    on_target_delivered=on_target_delivered,
-                    history_group_id=history_group_id,
-                    history_source=history_source,
-                )
-                if target_interval > 0:
-                    await asyncio.sleep(target_interval)
-            elif record_merge_placeholders:
-                for batch in batches:
-                    result.pushes.append(
-                        ScheduledPushResult(
-                            username=batch.username,
-                            new_count=len(batch.tweets),
-                            success_targets=0,
-                            total_targets=0,
-                            batch_key=self._scheduled_batch_key(batch),
-                        )
-                    )
-            await self._send_merged_updates(
-                batches,
-                result,
-                merge_targets,
-                target_interval,
-                group_label=group_label,
-                batch_summary=batch_summary,
-                on_target_delivered=on_target_delivered,
-                history_group_id=history_group_id,
-                history_source=history_source,
-            )
-            return
-
-        result.push_mode = "per_user"
-        await self._send_per_user_updates(
-            batches,
-            result,
-            ordinary_targets,
-            target_interval,
-            user_interval,
-            merge_existing_stats=merge_existing_stats,
-            group_label=group_label,
-            batch_summary=batch_summary,
-            on_target_delivered=on_target_delivered,
-            history_group_id=history_group_id,
-            history_source=history_source,
-        )
-
-    @staticmethod
-    async def _mark_batch_target_delivered(
-        batch: PendingTweetBatch,
-        target: str,
-        on_target_delivered=None,
-    ) -> None:
-        if on_target_delivered is None:
-            batch.delivered_targets.add(target)
-            return
-        result = on_target_delivered(batch, target)
-        if inspect.isawaitable(result):
-            await result
-        batch.delivered_targets.add(target)
-
-    @classmethod
-    def _delivery_history_status_ids(cls, outcome, tweets) -> tuple[str, ...]:
-        """Return only tweet IDs confirmed delivered by this outcome."""
-        available = tuple(
-            str(getattr(tweet, "status_id", "") or "")
-            for tweet in tweets
-            if str(getattr(tweet, "status_id", "") or "")
-        )
-        if cls._delivery_is_complete(outcome):
-            return available
-        status = str(
-            getattr(outcome, "delivery_status", "") or ""
-        ).strip().lower()
-        if status != "partial_failed":
-            return ()
-        delivered = {
-            str(status_id or "")
-            for status_id in getattr(outcome, "delivered_status_ids", ()) or ()
-            if str(status_id or "")
-        }
-        return tuple(status_id for status_id in available if status_id in delivered)
-
-    @staticmethod
-    def _delivery_is_complete(outcome) -> bool:
-        """Return whether an outcome is safe to advance the shared seen key."""
-        return bool(getattr(outcome, "success", False)) and str(
-            getattr(outcome, "delivery_status", "success") or "success"
-        ).strip().lower() != "failed"
-
-    @staticmethod
-    def _all_targets_delivered(
-        targets: list[str], batch: PendingTweetBatch
-    ) -> bool:
-        """Reject vacuous success when a delivery path has no targets."""
-        return bool(targets) and set(targets).issubset(batch.delivered_targets)
-
-    def _merge_tweet_threshold(self) -> int:
-        return configured_merge_tweet_threshold(self.config)
-
-    def _media_only_effective(self, group: ScheduleGroup) -> bool:
-        return bool(
-            group.media_only_enabled and not self._media_only_unavailable_reason(group)
-        )
-
-    async def _handle_media_prepare_status(
-        self,
-        result: ScheduledCheckResult,
-        batch: PendingTweetBatch,
-        status: str,
-        error: str,
-        seen_map: dict[str, list[str]],
-        group_id: str,
-    ) -> None:
-        batch.media_status = status
-        await self._cleanup_batch_media(batch)
-        status_id = str(batch.tweets[0].status_id or "") if batch.tweets else ""
-        if status == MEDIA_STATUS_POLICY_SKIPPED:
-            result.media_only_skipped += 1
-            if status_id:
-                await self._store_incremental_seen_ids(
-                    group_id,
-                    batch.username,
-                    [status_id],
-                    seen_map,
-                )
-            self._log_verbose_info(
-                "[NitterTweets] 仅媒体推文按全局媒体策略跳过: "
-                f"username={batch.username}, status={status_id}"
-            )
-            return
-
-        result.media_only_retrying += 1
-        logger.warning(
-            "[NitterTweets] 仅媒体推文暂未准备好，下轮重试: "
-            f"username={batch.username}, status={status_id}, "
-            f"status={status}, error={error or 'no media'}"
-        )
-
-    @staticmethod
-    def _has_prepared_media(tweets: list[TweetItem]) -> bool:
-        return any(
-            media.path and (media.is_image or media.is_video)
-            for tweet in tweets
-            for media in getattr(tweet, "media", [])
-        )
-
-    def _media_only_unavailable_reason(self, group: ScheduleGroup) -> str:
-        return media_only_unavailable_reason(
-            self.config,
-            media_only_enabled=group.media_only_enabled,
-        )
-
-    @staticmethod
-    def _should_merge_batches(batches: list[PendingTweetBatch], threshold: int) -> bool:
-        if threshold <= 0:
-            return False
-        return sum(len(batch.tweets) for batch in batches) >= threshold
-
-    def _split_immediate_targets(
-        self, targets: list[str], merge_threshold: int
-    ) -> tuple[list[str], list[str]]:
-        if merge_threshold <= 0:
-            return targets, []
-        merge_targets, ordinary_targets = self._split_merge_targets(targets)
-        if not merge_targets:
-            return targets, []
-        return ordinary_targets, merge_targets
-
-    @staticmethod
-    def _dedupe_targets(targets: list[str]) -> list[str]:
-        result: list[str] = []
-        seen: set[str] = set()
-        for target in targets:
-            if not target or target in seen:
-                continue
-            seen.add(target)
-            result.append(target)
-        return result
-
-    def _split_merge_targets(self, targets: list[str]) -> tuple[list[str], list[str]]:
-        merge_targets = []
-        ordinary_targets = []
-        for umo in targets:
-            if self.sender.supports_merged_forward_for_umo(self.context, umo):
-                merge_targets.append(umo)
-            else:
-                ordinary_targets.append(umo)
-        return merge_targets, ordinary_targets
-
-    def _order_targets_for_push(self, targets: list[str]) -> list[str]:
-        ordinary_targets = []
-        merge_targets = []
-        for umo in targets:
-            if self.sender.supports_merged_forward_for_umo(self.context, umo):
-                merge_targets.append(umo)
-            else:
-                ordinary_targets.append(umo)
-        return ordinary_targets + merge_targets
-
-    @staticmethod
-    def _tweet_batches(batches: list[PendingTweetBatch]) -> list[tuple[str, str, list]]:
-        return [(batch.username, batch.instance, batch.tweets) for batch in batches]
-
-    @staticmethod
-    def _format_push_batch_summary(
-        batches: list[PendingTweetBatch],
-        group_label: str,
-        action_text: str,
-    ) -> str:
-        if not batches:
-            return ""
-        return TweetMessageRenderer.format_batch_summary(
-            NitterTweetScheduler._tweet_batches(batches),
-            group_label=group_label,
-            action_text=action_text,
-        )
-
-    @staticmethod
-    def _append_fetch_failure_summary(
-        summary: str,
-        failed_users: dict[str, str],
-    ) -> str:
-        if not failed_users:
-            return summary
-        failed_items = []
-        for user, error in failed_users.items():
-            label = str(user)
-            if label.startswith("q:"):
-                failed_items.append(f"{label}: {error}")
-            else:
-                failed_items.append(f"@{label}: {error}")
-        failure_summary = "抓取失败：" + scheduler_format_limited_values(
-            failed_items,
-            limit=5,
-            separator="；",
-        )
-        if summary.strip():
-            return f"{summary.rstrip()}\n{failure_summary}"
-        return failure_summary
-
-    @staticmethod
-    def _push_group_label(group: ScheduleGroup) -> str:
-        if is_default_group(group.group_id):
-            return DEFAULT_GROUP_NAME
-        return str(group.name or group.group_id).strip() or group.group_id
-
-    async def _store_incremental_seen_ids(
-        self,
-        group_id: str,
-        username: str,
-        status_ids: list[str],
-        seen_map: dict[str, list[str]],
-    ) -> None:
-        ids = [str(item) for item in status_ids if item]
-        if not ids:
-            return
-        current = seen_map.get(username, [])
-        if not isinstance(current, list):
-            current = []
-        seen_map[username] = self._merge_seen_ids(ids, current)
-        await self.storage.add_seen_ids(group_id, username, ids)
-
-    @staticmethod
-    def _format_merge_threshold(threshold: int) -> str:
-        return scheduler_format_merge_threshold(threshold)
-
-    @staticmethod
-    def _format_group_schedule(group: ScheduleGroup) -> str:
-        return scheduler_format_group_schedule(group)
-
-    def _append_group_status(
-        self,
-        lines: list[str],
-        group: ScheduleGroup,
-        seen_count: int | None = None,
-    ) -> None:
-        type_label = "标签" if group.is_tag_group else "博主"
-        account_count = len(group.account_keys)
-        lines.append(
-            "- "
-            f"{group.name} ({group.group_id}): "
-            f"{'启用' if group.enabled else '关闭'}，"
-            f"类型 {type_label}，"
-            f"订阅 {account_count}，目标 {len(group.targets)}，"
-            f"{self._format_group_schedule(group)}"
-        )
-        if group.aliases:
-            lines.append("  别名: " + self._format_limited_values(group.aliases))
-        if group.is_tag_group:
-            lines.append(
-                "  搜索订阅: "
-                f"{len(group.queries)} 个（配置 {group.queries_info.raw_count} 项，"
-                f"重复 {len(group.queries_info.duplicates)} 项，"
-                f"无效 {len(group.queries_info.invalid_entries)} 项）"
-            )
-        else:
-            lines.append(
-                "  关注账号: "
-                f"{len(group.users)} 个（配置 {group.users_info.raw_count} 项，"
-                f"重复 {len(group.users_info.duplicates)} 项，"
-                f"无效 {len(group.users_info.invalid_entries)} 项）"
-            )
-        lines.append(
-            f"  推送目标: {len(group.targets)} 个"
-            f"（无效 {len(group.invalid_targets)} 个）"
-        )
-        if seen_count is not None:
-            lines.append(f"  已记录索引: {seen_count} 个")
-        daily_times = group.daily_check_times
-        if daily_times:
-            formatted_times = ", ".join(
-                f"{hour:02d}:{minute:02d}" for hour, minute in daily_times
-            )
-            lines.append(f"  每日时间: {formatted_times}")
-        if group.is_tag_group and group.queries:
-            labels = [f"{item.query} ({item.type})" for item in group.queries]
-            lines.append("  订阅查询: " + self._format_limited_values(labels))
-            if group.queries_info.duplicates:
-                lines.append(
-                    "  重复查询: "
-                    + self._format_limited_values(group.queries_info.duplicates)
-                )
-            if group.queries_info.invalid_entries:
-                lines.append(
-                    "  无效查询: "
-                    + self._format_limited_values(group.queries_info.invalid_entries)
-                )
-        elif group.users:
-            usernames = [f"@{username}" for username in group.users]
-            lines.append("  订阅账号: " + self._format_limited_values(usernames))
-            if group.users_info.duplicates:
-                lines.append(
-                    "  重复订阅: "
-                    + self._format_limited_values(group.users_info.duplicates)
-                )
-            if group.users_info.invalid_entries:
-                lines.append(
-                    "  无效订阅: "
-                    + self._format_limited_values(group.users_info.invalid_entries)
-                )
-        if group.targets:
-            lines.append("  推送目标:")
-            for umo in group.targets[:8]:
-                lines.append(f"  - {umo}")
-            if len(group.targets) > 8:
-                lines.append(f"  - ... 还有 {len(group.targets) - 8} 个")
-        if group.invalid_targets:
-            lines.append(
-                "  无效目标: " + self._format_limited_values(group.invalid_targets)
-            )
-
-    @staticmethod
-    def _format_limited_values(values: list[str], limit: int = 10) -> str:
-        return scheduler_format_limited_values(values, limit=limit)
-
-    @staticmethod
-    def _format_daily_times(times: list[tuple[int, int]]) -> str:
-        return scheduler_format_daily_times(times)
-
-    @staticmethod
-    def _format_next_daily_time(times: list[tuple[int, int]]) -> str:
-        return scheduler_format_next_daily_time(times)
-
-    @staticmethod
-    def _format_timestamp(timestamp: int) -> str:
-        return scheduler_format_timestamp(timestamp)
-
-    async def _get_seen_map(
-        self, group_id: str = GLOBAL_GROUP_ID
-    ) -> dict[str, list[str]]:
-        return await self.storage.get_group_seen_map(group_id)
-
-    async def _get_scan_watermarks(
-        self,
-        group_id: str,
-        seen_map: dict[str, list[str]],
-    ) -> dict[str, list[str]]:
-        getter = getattr(self.storage, "get_group_scan_watermarks", None)
-        if callable(getter):
-            stored_watermarks = await getter(group_id)
-            # Older databases and lightweight test adapters may contain seen
-            # IDs but no dedicated anchor row yet. Infer only the missing
-            # entries; an explicit empty anchor window means the source was
-            # initialized with no historical status ID and must be preserved.
-            for username, status_ids in seen_map.items():
-                if username in stored_watermarks:
-                    continue
-                if not isinstance(status_ids, list):
-                    continue
-                if status_ids:
-                    stored_watermarks[username] = list(status_ids[:20])
-            return stored_watermarks
-
-        # Keep older storage fakes and external adapters usable while they add
-        # the dedicated scan-anchor API.
-        return {
-            username: list(status_ids[:20])
-            for username, status_ids in seen_map.items()
-            if isinstance(status_ids, list) and status_ids
-        }
-
-    async def _set_scan_watermark(
-        self,
-        group_id: str,
-        username: str,
-        status_ids: list[str] | str | None,
-    ) -> None:
-        setter = getattr(self.storage, "set_scan_watermark", None)
-        if callable(setter):
-            await setter(group_id, username, status_ids)
-
-    async def _put_seen_map(
-        self, group_id: str, seen_map: dict[str, list[str]]
-    ) -> None:
-        await self.storage.put_group_seen_map(group_id, seen_map)
-
-    def _merge_seen_ids(self, new_ids: list[str], old_ids: list[str]) -> list[str]:
-        return self.storage.merge_seen_ids(new_ids, old_ids)
-
-    @classmethod
-    def _select_new_tweets_after_scan_watermark(
-        cls,
-        tweets: list[TweetItem],
-        seen_ids: list[str],
-        watermark_ids: list[str] | None,
-    ) -> tuple[list[TweetItem], list[str]]:
-        del watermark_ids
-        seen_set = set(str(item) for item in seen_ids)
-        new_tweets: list[TweetItem] = []
-
-        for tweet in tweets:
-            status_id = str(tweet.status_id or "")
-            if not status_id or status_id in seen_set:
-                continue
-            new_tweets.append(tweet)
-
-        return new_tweets, []
-
-    @classmethod
-    def _max_numeric_status_id(cls, seen_ids: list[str]) -> int | None:
-        numeric_ids = [
-            status_number
-            for status_id in seen_ids
-            if (status_number := cls._parse_numeric_status_id(str(status_id)))
-            is not None
-        ]
-        if not numeric_ids:
-            return None
-        return max(numeric_ids)
-
-    @staticmethod
-    def _parse_numeric_status_id(status_id: str) -> int | None:
-        value = str(status_id or "").strip()
-        if not value or not value.isdigit():
-            return None
-        return int(value)
 
     def _watch_users(self) -> list[str]:
         return self.config_reader.watch_users()
